+++ resources/views/bootstrap-4/form-range.blade.php
@@ -1,9 +1,15 @@
 <div class="form-group">
     <x-form-label :label="$label" :for="$attributes->get('id') ?: $id()" />
 
-    <textarea
+    <input
+        {!! $attributes->merge(['class' => 'form-control-range' . ($hasError($name) ? ' is-invalid' : '')]) !!}
+
+        type="range"
+
         @if($isWired())
             wire:model{!! $wireModifier() !!}="{{ $name }}"
+        @else
+            value="{{ $value }}"
         @endif
 
         name="{{ $name }}"
@@ -11,9 +17,7 @@
         @if($label && !$attributes->get('id'))
             id="{{ $id() }}"
         @endif
-
-        {!! $attributes->merge(['class' => 'form-control ' . ($hasError($name) ? 'is-invalid' : '')]) !!}
-    >@unless($isWired()){!! $value !!}@endunless</textarea>
+    />
 
     {!! $help ?? null !!}