+++ resources/views/bootstrap-5/form-checkbox.blade.php
@@ -1,6 +1,9 @@
-<div class="form-check">
-    <input {!! $attributes->merge(['class' => 'form-check-input ' . ($hasError($name) ? 'is-invalid' : '')]) !!}
+<div class="form-check @if(null !== $attributes->get('switch')) form-switch @endif @if(null !== $attributes->get('inline')) form-check-inline @endif">
+    <input
+        {!! $attributes->merge(['class' => 'form-check-input' . ($hasError($name) ? ' is-invalid' : '')]) !!}
+
         type="checkbox"
+
         value="{{ $value }}"
 
         @if($isWired())