+++ resources/views/tailwind/form-checkbox.blade.php
@@ -6,9 +6,9 @@
 
             @if($isWired())
                 wire:model{!! $wireModifier() !!}="{{ $name }}"
-            @else
-                name="{{ $name }}"
             @endif
+
+            name="{{ $name }}"
 
             @if($checked)
                 checked="checked"