+++ src/Components/FormRange.php
@@ -2,15 +2,13 @@
 
 namespace ProtoneMedia\LaravelFormComponents\Components;
 
-class FormInput extends Component
+class FormRange extends Component
 {
     use HandlesValidationErrors;
     use HandlesDefaultAndOldValue;
 
     public string $name;
     public string $label;
-    public string $type;
-    public bool $floating;
 
     public $value;
 
@@ -22,18 +20,14 @@ class FormInput extends Component
     public function __construct(
         string $name,
         string $label = '',
-        string $type = 'text',
         $bind = null,
         $default = null,
         $language = null,
-        bool $showErrors = true,
-        bool $floating = false
+        bool $showErrors = true
     ) {
         $this->name       = $name;
         $this->label      = $label;
-        $this->type       = $type;
         $this->showErrors = $showErrors;
-        $this->floating   = $floating && $type !== 'hidden';
 
         if ($language) {
             $this->name = "{$name}[{$language}]";