+++ tests/Feature/views/bootstrap-input-group.blade.php
@@ -1,0 +1,7 @@
+<x-form>
+    <x-form-input-group>
+        <x-form-input name="name" id="name" />
+        <x-form-input-group-text>@</x-form-input-group-text>
+        <x-form-input name="username" id="username"></x-form-input>
+    </x-form-input-group>
+</x-form>